migrate code editor helpers to individual files.
