+++ AvalonStudio/AvalonStudio.Languages.Xaml/TerminateElementCodeEditorHelper.cs
@@ -35,39 +35,19 @@ namespace AvalonStudio.Languages.Xaml
         }
     }
 
-    class InsertQuotesForPropertyValueCodeEditorHelper : ICodeEditorInputHelper
-    {
-        public void AfterTextInput(ILanguageService languageServivce, ITextDocument document, TextInputEventArgs args)
-        {
-            if (args.Text == "=")
-            {
-                var textBefore = document.Text.Substring(0, Math.Max(0, document.Caret - 1));
-                if (textBefore.Length > 2 && textBefore[textBefore.Length - 1] != '/')
-                {
-                    var state = XmlParser.Parse(textBefore);
-                    if (state.State == XmlParser.ParserState.StartAttribute)
-                    {
-                        var caret = document.Caret;
-                        document.Replace(caret, 0, "\"\" ");
-                        document.Caret = caret + 1;
-                    }
-                }
-            }
-        }
-
-        public void BeforeTextInput(ILanguageService languageService, ITextDocument document, TextInputEventArgs args)
-        {
-        }
-    }
-
     class XamlIndentationCodeEditorHelper : ICodeEditorInputHelper
     {
         public void AfterTextInput(ILanguageService languageServivce, ITextDocument document, TextInputEventArgs args)
         {
+            
+        }
+
+        public void BeforeTextInput(ILanguageService languageService, ITextDocument document, TextInputEventArgs args)
+        {
             if (args.Text == "\n")
             {
                 //Check if we are not inside a tag
-                var textBefore = document.Text.Substring(0, Math.Max(0, document.Caret - 1));
+                var textBefore = document.Text.Substring(0, Math.Max(0, document.Caret));
                 var state = XmlParser.Parse(textBefore);
                 if (state.State == XmlParser.ParserState.None)
                 {
@@ -75,40 +55,15 @@ namespace AvalonStudio.Languages.Xaml
                     var idx = textBefore.LastIndexOf('>');
                     if (idx != -1)
                     {
-                        state = XmlParser.Parse(textBefore.Substring(0, Math.Max(0, idx - 1)));
-                        if (state.TagName.StartsWith('/'))
+                        state = XmlParser.Parse(textBefore.Substring(0, Math.Max(0, idx)));
+
+                        if(state.State == XmlParser.ParserState.StartElement && document.Text[document.Caret] == '<')
                         {
-                            //TODO: find matching starting tag. XmlParser can't do that right now.
-                            return;
-                        }
-                        //Find starting '<'
-                        bool insideAttribute = false;
-                        for (; idx >= 0; idx--)
-                        {
-                            var ch = textBefore[idx];
-                            if (ch == '"')
-                                insideAttribute = !insideAttribute;
-                            if (ch == '<' && !insideAttribute)
-                            {
-                                var textBeforeTag = textBefore.Substring(0, idx);
-                                var lineStartIdx = textBeforeTag.LastIndexOf('\n');
-                                if (lineStartIdx != -1)
-                                {
-                                    //TODO: Do something about '\t' characters
-                                    var prefixLength = (idx - lineStartIdx) - 1;
-                                    document.Replace(document.Caret, 0,
-                                        new string(' ', prefixLength));
-                                }
-                                return;
-                            }
+                            document.Replace(document.Caret, 0, "\n");
                         }
                     }
                 }
             }
-        }
-
-        public void BeforeTextInput(ILanguageService languageService, ITextDocument document, TextInputEventArgs args)
-        {
         }
     }
 }
